Fix and move test
