+++ tests/test_log_metric.py
@@ -4,6 +4,18 @@ import numpy as np
 import pytest
 
 from dvclive import Live
+from dvclive.error import InvalidDataTypeError
+
+
+@pytest.mark.parametrize("invalid_type", [{0: 1}, [0, 1], (0, 1)])
+def test_invalid_metric_type(tmp_dir, invalid_type):
+    dvclive = Live()
+
+    with pytest.raises(
+        InvalidDataTypeError,
+        match=f"Data 'm' has not supported type {type(invalid_type)}",
+    ):
+        dvclive.log_metric("m", invalid_type)
 
 
 @pytest.mark.parametrize(
